freeze googletest version
We have googletest as a dependency for our tests. This dependency is
defined by cmake and fetched when we run make, but we were always
fetching master, which means if there is a breaking change upstream our
build will break without notice (which happened recently). To avoid
this, we freeze the googletest version to the last stable, and once
there's a new googletest release we can upgrade it here (making changes
to our code if necessary).

Fixes: https://github.com/iovisor/bpftrace/issues/122