clang-parser: Extend support for char type
This extends support for the char type by making the clang
parser recognize CXType_Char_U as well.

Currently, the clang parser assumes that the char type is
always implemented as a signed char. However, for certain
architectures like powerpc64, the char type is unsigned
and the parser fails to recognize such variables.

Signed-off-by: Sandipan Das <sandipan@linux.ibm.com>