Fix problems with zero()'ing maps
This change ensures that any maps using the count(), sum(), avg(),
min(), max(), stats(), hist() or lhist() builtins can be zero'd without
hanging the bpftrace process or with generating garbage data in the map.

Special attention is drawn to the changes here in the min_value()
function. I have changed it so that it what you'd expect with signed
quantities. However, I feel the propoer fix would be to change the BPF
generation for the min() function so that it works with 64 bit values
correctly and remove the need for conversion of the stored value.