usdt: path is optional when attaching to pid
When instrumenting a pid the path to the binary is optional, since the probes
can be determined by discovering the probes by attaching to the process.

Practically, the path is resolved and associated with the attach point during
the semantic analyzer path, though the irbuilder must also attach to the
process to resolve the arguments.

Remove GetProviderFromPath in favor of discovering the provider by attaching to
the process, which is the more reliable method. The path only has a passing
relationship to the provider as defined by stapsdt.

Upstream `bcc_usdt_new_frompid` has an issue when instrumenting a process that
is in a mount namespace, so for now don't include the path when attaching to a
pid.

Relatedly, the `stat()` that occurs in the semantic analyzer is removed since
it isn't namespace aware.