build: define an option to not use bundled dependencies
By default, bpftrace build system will fetch libbcc and gtest, compile
them and link to the result. Distributions need to use existing
package and not fetch stuff from Internet. In Debian, libbcc is
available as libbpfcc-dev and gtest is available as libgtest-dev and
libgmock-dev.

We define three new options:
 - USE_BUNDLED,
 - USE_BUNDLED_BCC, and
 - USE_BUNDLED_GTEST

They are OFF by default, so for regular users, there is no change.
When the first one is set to OFF, the two other ones are set to OFF
too (unless explicitely set to ON).

Best practice in CMake recommends to use `find_package()` and ship
`FindXXX.cmake` files to do the hardwork. I think it would be better
for libbcc to ship the right `FindXXX.cmake` instead of defining it
here. In the meantime, it's defined here. Currently, libbpfcc-dev
package doesn't ship the .pc file, so I am using `find_path()` and
`find_library()` to locate the appropriate paths. I have asked
libbpfcc maintainer to include the .pc file.

As for gtest, CMake ships with `FindGTest()`, so we use that. But
there is no `FindGMock()`. There are some versions online, but I've
just used `find_path()` and `find_library()`.

Currently, the result doesn't work because bpftrace is using some
include files not shipped with libbcc:

 - `common.h`
 - `syms.h` (then `ns_guard.h`)
 - `frontends/clang/kbuild_helper.h`

I don't know if this is a problem with bpftrace which shouldn't use
these includes or with libbcc which should ship them. If I copy them
manually, it works. Tests are also succeeding (as root only, something
that will cause problems for distributions too, but I'll look at that
later).