[bcc] fix for new bpf_attach_kprobe signature
bcc changed the signature of bpf_attach_kprobe to accept a new argument:
`maxactive`, which determines the maximum limit of how many calls in
parallel a kretprobe attached to a kernel function can catch. The new
signature is not backwards compatible, thus our build broke. Since
there's no way to feature detect the existance of an argument in a
function using cmake, we use a function pointer with the long signature
and cast bpf_attach_kprobe to it. If we're on an older bcc version, the
last argument will be allocated in a register (according to the
platform's calling convention), but should be ignored inside the
function.

Ref: https://github.com/torvalds/linux/commit/696ced4fb1d76802f864d8848aa471663
Ref: https://github.com/iovisor/bcc/pull/2224
Fixes: https://github.com/iovisor/bpftrace/issues/453