add basic support to enums (treat them as constants)
This adds basic support for C-style enumerators. By basic support I mean
we are treating them as constants (almost #defines) instead of new
types. Maybe in the future we want to add more sophisticated support for
them (allowing cast to an enum type, checking if an assignment is within
the enum range, etc.). Maybe we will want to keep it simple. But for now
this should solve the most common use cases, especially when dealing
with enumerators from Kernel headers.

Ref: #520