[clang_parser] fix struct definition from headers
Regression caused by iovisor/bpftrace@80ce138c7. With the changes on how
we identify the parent struct, we ended up with our parent cursor in the
header file sometimes instead of a valid cursor. This PR checks if the
parent cursor is a struct, otherwise returns an empty string and let the
caller handle the situation (which is similar to the previous behavior).