[tests] don't check how many structs clang parsed
Since we have implicit imports from the Kernel, there's no way to
reliabley tell how many structs clang was supposed to parse. As a
result, tests could fail in some environments and succeed in others.
This commit removes the check for how many structs clang parsed in our
tests. This can be added again in the future if we make clang_parser
more decoupled (test without the kernel flags, for example) or if we
only retain structures used by our bpf program (which would be ideal).

Fixes: https://github.com/iovisor/bpftrace/issues/569