Allow non-zero lhist min value
Signed division isn't allowed when using bpf as a target[1]. When `min = 0`
the llvm can apparently figure out that signed division isn't required
and it rewrites it to unsigned. However when `min > 0` the  llvm seems
to account for `value - min < 0` and uses signed division.
As `value < min` has already been checked earlier we can be sure that
`value - min >= 0` and always use unsigned division.

[1]: https://github.com/llvm-mirror/llvm/commit/2462cfbd350cbabad4c33c8aeb7601a7219be2fd