tools: Fix header import order of tcp tools
There's a limitation in the current implementation for imports which
means that structures in second or subsequent #include statements are
not able to be used for casting within functions. This causes tcpaccept
(and similar programs) to fail with:

    $ ./tcpaccept.bt
    ...
    Unknown struct/union: 'sock'

Fix this by reordering the includes to import the header with the sock
definition first.