Parse negative integer literals correctly
The parser was parsing "-10" an a unary op on an integer. This
led to reading invalid memory because the semantic analyser
did a static_cast upcast to `Integer` from `Expression`
(which was actually an `Unop`).

This patch requires the constant lhist arguments to be positive
integer literals. This patch also modifies the parser so it correctly
parses negative inteters as literals (vs `Unop`s).

Test Plan:
Before:
```
$ sudo ./build/src/bpftrace -e 'i:ms:1 { @=lhist(-10, -10, 10, 1); }'
lhist() min must be less than max (provided min 13989904 and max 10)
lhist() step is too large for the given range (provided step 1 for range
-13989894)
```

After:
```
$ sudo ./build/src/bpftrace -e 'i:ms:1 { @=lhist(-10, -10, 10, 1); }'
lhist() min must be non-negative (provided min -10)
```

This closes #618.