Add option to redirect trace output to file
This adds the option to redirect all output generated by the "trace
program" to be output to a file as discussed in #605. All other output
generated by `bpftrace` will still be sent either `stdout` or `stderr`.

It's implementation is quite straight forward. The `BPFtrace` class
stores a reference to an `ostream` object which will be used for all
output, replacing `cout` everywhere.
The `printf` command now uses `snprintf` to print to a buffer which then
gets written to the output stream. As the size isn't known initially
`sprintf` is called with a fixed size stack buffer first. If the buffer
turned out to be to small a new buffer of the required size will be
allocated (`snprintf` returns the required size).

Instead of calling `system` the `system` builtin now uses `exec_system`
from as that captures `stdout`. `stderr` is not captured (yet).

Basic tests are added too. As the tests suite contains no temp file
support the command was extended with an final `rm` to delete the temp
file. Which feels like a better solution than adding special commands to
the test suite for 4 tests.