Limit the amount of byte `cat` can read and output
The `cat` command currently doesn't check file sizes before printing. It
will happily print a 1GiB file to the terminal and kill your system
with it.

To prevent that from happening this introduces a new environment var,
`BPFTRACE_CAT_BYTES_MAX`, which limits the amount of bytes read by the
`cat` builtin.

```
$ sudo BPFTRACE_CAT_BYTES_MAX=1 ./src/bpftrace -e 'i:ms:1 { cat("/proc/uptime"); exit(); }'
Attaching 1 probe...
2
$ sudo BPFTRACE_CAT_BYTES_MAX=4 ./src/bpftrace -e 'i:ms:1 { cat("/proc/uptime"); exit(); }'
Attaching 1 probe...
2758
$ sudo BPFTRACE_CAT_BYTES_MAX=512 ./src/bpftrace -e 'i:ms:1 { cat("/proc/uptime"); exit(); }'
Attaching 1 probe...
27596.19 53430.89
```

Internally it reserves a single 4k buffer on the stack and then reads
and outputs the file in blocks of 4k. It will stop when either the limit
or EOF is hit.
No effort is done to calculate the file size to limit the buffer as
"kernel fileystems", which seem like the main use case for `cat`, do not
supply have a predetermined size.
If an error is encounted during the read a generic error will be printed
and the command will return.