Fix some bugs with positional parameters
This invalid script used to compile:
    bpftrace -e 'BEGIN { str(0); @x = $1 }' a
Now bpftrace correctly recognises that $1 is not inside the str() call

This script did not function correctly since "234" was
treated as a numeric value despite being inside str():
    bpftrace -e 'BEGIN { @x = str($1) }' 234
I've just disabled using numeric values in str() because the
codegen is not set up for it, although this is something that
should be valid.

Positional parameters are no longer marked as literals because we're
not set up for this and it resulted some dodgy memory accesses. We
probably do want them to be capable of being used as literals, but
more work is required to support that.