Fix k[ret]probe_order runtime tests
On my system, I was getting

  Attaching kprobe:sys_execve
  cannot attach kprobe, probe entry may not exist
  Error attaching probe: 'kprobe:sys_execve'

It looks like commit e145242ea0df6b7d2 in the kernel changed
how syscall stubs are named. That was about a year ago so it
makes sense that now we're seeing some of that newer code being
run around the world.

This patch moves the kprobe rutime tests to use a more stable
kernel interface.