Fix exiting prematurely on strace attach
When strace is attached to bpftrace, epoll_wait returns EINTR, which
caused us to exit pre-maturely. However, an SIGINT also has the same
behavior.

To get around this, we store a global which will flip to true when a
SIGINT is called, and use that to check if we should continue or exit.

Closes #712