Remove SIGINT handler after bpf program is finished running
Normally, we suppress the default handler of SIGINT to instead exit
the bpf program. Once the bpf program is done running, we may have
extended post processing, and it would be ideal to exit on SIGINT
during this. This patch re-enables the default SIGINT handler prior to
post-processing to get this behavior.

Closes #713