utils: unpack kheaders.tar.xz if necessary
When CONFIG_IKHEADERS is enabled, a tarball containing the kernel
headers is available in /sys/kernel, allowing the kernel type
definitions to be used even if the headers have not been separately
distributed to /lib on the target.

This is akin to [how bcc handles it][1], but with the new path in /sys
after [2].  Notably, use the same path in /tmp, so that the tarball will
be unpacked only once between the two tools.

[1]: https://github.com/iovisor/bcc/commit/ae92f3ddb6aa5b81c750abf3540b99f24d219e67
[2]: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=f7b101d33046a837c2aa4526cef28a3c785d7af2