Add new environment variable BPFTRACE_LOG_SIZE (#751)
Per Issue #584, running with -v can cause programs to fail, due to BPF_LOG_SIZE not being large enough to store the program. As noted in the issue, log size should be settable from an environment variable and bpftrace should give a better error message when this occurs.

This PR introduces a new environment variable BPFTRACE_LOG_SIZE as a replacement of the const int BPF_LOG_SIZE = 400 * 1024; in attached_probe.cpp. Additionally we add logging as a hint to the user to increase BPFTRACE_LOG_SIZE when ENOSPC is received. Finally we've added runtime tests for verification.