Add source location info to the parser error printer
With this change the error printer gets access to the source code which
can be used to improve error messages, e.g.:

```
../opensnoop.bt:23:17-25: ERROR: Type mismatch for '<': comparing 'integer' with 'string'
        $ret =  0 < "aa"; // work around or
                ~~~~~~~~

../opensnoop.bt:24:2-10: ERROR: Type mismatch for '<': comparing 'integer' with 'string'
    0 < "aa";
    ~~~~~~~~
```

For this to work the error printer needs a way to access the relevant
source code. As bpftrace is built for small single file scripts we can
safely store the whole program source in memory without having to worry
about memory consumption. This keeps the code a bit simpler as dealing
with files can be avoided. It also keeps the code path for scripts passed
on stdin and through a file the same.
With the whole source stored in the bpftrace object the driver interface
can be refactored a bit. Instead of a single `parse_X` call it is now
split into a source call which reads the source and stores it, and a
parse call which parses the previously stored source.

bpftrace scripts allow for a mix of tabs and spaces in the scripts. This
is an issue for the marker in the error format as the width of tabs
depends on the terminal. The tab is one character but could be any
number of spaces when displayed, so we do not know how many `~`s to
print in the marker. Which leads to output like:

```
test.bt:1:8-10: ERROR: syntax error, unexpected variable, expecting {
i:s:1        $a
       ~~
```

The fix chosen here is to print a tab as 4 spaces. This means the column
number displayed won't exactly print the printed output. But it sill
matches the column number in the source file, which is what maters.