Add resolve_offset_uprobe functions
To properly resolve offset for uprobe, which can be specified
now with virtual address or symbol + offset.

Keeping the original offset() code in resolve_offset function,
which is still used for usdt probes.

Adding 'AttachedProbe::offset_' variable to carry the resolved
offset value.

The reason for resolve_offset_uprobe function is the
extra checking/resolving on for uprobes we need to have
plus the offset check coming in following patches.