ntop: fix alignment issues
ntop would behave unpredictably when used with if/else statements for
differently-sized ntop buffers (8/20, depending on inet AF type).
Upcasting to 20 fixes the issue for some tools, but caused other tools
to stop translating the address correctly. This second issue was caused
by LLVM adding pad between elements of the printf buffer, which caused
our printf implementation to read from incorrect memory addresses. To
prevent LLVM from adding paddings, we can change the size of ntop to a
stack-aligned size. 24 is the lowest stack-aligned size, so we use 24.

Fixes: https://github.com/iovisor/bpftrace/issues/809