Do not emit GEP instruction when pushing string literals to stack (#667)
* Do not emit GEP instruction when pushing string literals to stack

I am able to crash bpftrace on one of asserts in llvm/Support/Casting.h when
building with LLVM8 and asserts enabled. Minimal repro I found is:
```
bpftrace -e 'BEGIN { printf("%s", "a"); }'
```
The crash occurs while creating GEP instruction when generating code pushing "a"
string literal to stack (see
https://github.com/iovisor/bpftrace/blob/master/src/ast/codegen_llvm.cpp#L46).
GEP expects its argument to be a `Value` of type `PointerType`, whereas type of
`Constant` pointed to by `const_str` is `ArrayType`. Proposed fix is to skip GEP
in this case, store opcode can handle arrays.

The issue might be hard to spot because with asserts disabled code seem to
work fine (fields of PointerType and ArrayType fortunately overlap in a way
that makes sense, `PointeeTy` in `PointerType` overlaps with `ContainedType`
added in `ArrayType`'s superclass).

Test plan:
Build bpftrace with LLVM8 and asserts enabled and make sure that I can't repro
the crash anymore. In additoin pick  run:
- biolatency.bt
- biosnoop.bt
- runqlat.bt
- tcpconnect.bt

* Fix assignment of printf_id_ in codegen for printf.

We want to store to 0th field of printf_args, we need to emit GEP.

* Fix perfoutput_func_type

Second argument is of type i64 (result of calling llvm.bpf.pseudo), 4th argument
is a structure that is defined dynamically, so just use type of the passed
data Value.

With this change `bpftrace -e 'BEGIN { printf("%s", "a"); }'` does not crash
anymore when asserts are enabled in llvm, clang, bcc and bpftrace, but there
are more type correcness issues to be fixed.