Add BPFTRACE_NO_USER_SYMBOLS env var (#800)
This env var lets the user disable user symbol resolution. We need this
because of iovisor/bcc#2421 , where symbol resolution can take too much
memory. I'll work on getting the proper fix in bcc.

The reason this isn't exposed as another stack mode is because this is
more of a workaround and not something we should expect the language
to support for a long time.

This patch also deletes an ambiguous (and possibly outdated) comment.

Test plan:
```
$ sudo BPFTRACE_NO_USER_SYMBOLS=1 ./build/src/bpftrace -e 'kprobe:do_sys_open{ @[ustack] = count(); }'
Attaching 1 probe...
^C

@[
    0x7f4cd9d5b626
]: 1
@[
    0x49bb6a
    0x492f00
    0x4bbdf5
    0x4b9bcf
    0x4b9ab6
    0x6b093b
    0x120b93b
    0x13eeb3b
    0x13f0859
    0x1244dd8
    0x1245147
    0x13f6132
    0x124415d
    0x1257526
    0x73bbd4
    0x76bb5a
    0x1255ec0
    0x73e94b
    0x73ac36
    0x4698c1
]: 1
@[
    0x7f4cd9d5b626
]: 1
@[
```

Then scroll up and down a bit and verify no symbols names are present.