Disable func, retval and reg for tracepoints
Unlike the `BPF_PROG_TYPE_KPROBE` probes the `BPF_PROG_TYPE_TRACEPOINT`
doesn't have access to the CPU register in `struct pt_regs`. Instead
they get the tracepoint fields passed.
Because of this the `func` and `retval` builtins and `reg` function
cannot be used for tracepoints as they depend on access to the `pt_regs`
struct.

[1]: Linux commit 98b5c2c65c2951772a8fc661f50d675e450e8bce:

> prior to the compilation of the program with exception that first 8
> bytes are reserved and not accessible to the program. This area is
> used to store the pointer to 'struct pt_regs' which some of the bpf
> helpers will use:
> +---------+
> | 8 bytes | hidden 'struct pt_regs *' (inaccessible to bpf program)
> +---------+
> | N bytes | static tracepoint fields defined in tracepoint/format (bpf readonly)
> +---------+
> | dynamic | __dynamic_array bytes of tracepoint (inaccessible to bpf yet)
> +---------+