Add signed types
bpftrace currently has a somewhat confusing mix of signed and unsigned
integers throughout the codebase. As an example, we can assign a
negative as map value but the printer will see it as unsigned[1].

This change introduces signed integer types by adding a new `is_signed`
attribute to the `SizedType`. This new attribute is implemented
throughout the codebase with the exception of the trace point parser,
which will be done in a future version.

With this change all integer literals are signed by default as negative
numbers are already supported. If there is a need for 64 bits unsigned
ints it should be easy to add. One option would be to use the `UL`
suffix that C uses too.
The way integers are printed has also been changed slightly. Unsigned
ints are now prefixed with the `unsigned` keyword, and their size in
bits is appended at the end.

This does not try to be smart about signed and unsigned binops, instead
if emits a warning when the mixing of signs can be an issue, roughly
following the C standard. Keeping it simple and leaving the conflict
resolution up to the user.

The Clang parser code has also been updated to handle the signs.

One ugly bit is the way map keys are compared. A common use for maps is
using them as lookup table for translating enums, like the `tcpdrop`
tools does. As integers are now signed by default this can lead to a
sign mismatch between the initialization and lookup of maps values. To
not break current script (not just those in the repo, but those of users
to) we skip the sign for mapkeys. Once we have a way to create unsigned
integer literals we should maybe revisit this and consider emitting a
warning.

[1]
```
sudo bpftrace_stable -e  'BEGIN { @[10]=-10; @[0]=10; @[5]=0; @[-10]=100; exit() }'

Program
 BEGIN
  =
   map: @
    int: 10
   int: -10
[..]
    call: exit

@[5]: 0
@[0]: 10
@[-10]: 100
@[10]: 18446744073709551606
```