Introduce int casts
Currently all integers are 64 bits internally. This causes issues when
dealing with smaller (bitsize) values returned by the kernel. An example
is:

```
kr:sock_recvmsg /retval < 0/{
  @[comm, retval]++;
}
```

`sock_recvmsg` returns a 32 bits int which, depending on the CPU
architecture, gets zero extended to 64 bits. Because of the zero
extension predicate will always be false.
With int casts the user can tell bpftrace that it should cast the value
before doing the comparison.

As this casts types to integers these new casts are of `Type::integer`
instead of `Type::cast`. This keeps the amount of modifications needed
to a minimum as the code already knows how to handle integers. The
codegen is only a single truncate instruction, other parts of the
codegen cast to the required width when required.

One special case is mapkey. If we set the `size` of the underlying type
to 8 then the cast itself will not be performed. The fix for this is to
insert an additional cast forcing a cast up to 64bits. This is done by
rewriting the AST.
An alternative would be to not force 64bit map keys but that requires
more code without clear benefit.