Make `stats` and `avg` signed
By forcing these functions to be signed by default we get proper
handling of negative input values. The downside is that we lose 1 bit
to the sign, so very large numbers might overflow.
If this turns out to be an issue we could look into proper sign
detection and related warnings when someone comes up with a use case.