Signed types (#834)
While working on #554 / #772 I realised that most of the work I was doing was adding signed integer support to the language so that the casts and printing make sense. My cast implementation is only a few lines of code. As int casts and signed ints are related but different things I decided to split this parts into a new MR.

Opening this now to get some feedback on the approach. So far it seems to work quite well with only minor modifications to  clang struct tests.

Example:

```
$ sudo bpftrace -e 'kr:sock_recvmsg { @=sum(-11); }'

@: -22

$ sudo bpftrace -e  'BEGIN { @[10]=-10; @[0]=10; @[5]=0; @[-10]=100; exit() }'

@[10]: -10
@[5]: 0
@[0]: 10
@[-10]: 100
```

----

Add signed types

bpftrace currently has a somewhat confusing mix of signed and unsinged
integers throughout the codebase. As an example, we can assign a
negative as map value but the printer will see it as unsigned[1].

This change introduces signed integer types by adding a new `is_signed`
attribute to the `SizedType` which is implemented throughout the
codebase with the exception of the trace point parser, which will be
done in a future version.

With this change all integer literals are signed by default as negative
numbers are already supported. If there is a need for 64 bits unsigned
ints it should be easy to add.

This does not try to be smart about signed and unsigned binops, instead
if prints a warning if the types do not match on both sides. I did not
want to introduce additional complexity without having a good reason to
do so.

[1]
```
sudo bpftrace_stable -e  'BEGIN { @[10]=-10; @[0]=10; @[5]=0; @[-10]=100; exit() }'

Program
 BEGIN
  =
   map: @
    int: 10
   int: -10
[..]
    call: exit

@[5]: 0
@[0]: 10
@[-10]: 100
@[10]: 18446744073709551606
```