Store the BPF map type in the map object
Throughout the codebase the `type` (integer, count etc) is used to make
assumptions about the BPF map type used, instead of having access to the
real type. This leads to bugs, an example:

in `map.cpp`
```
  if ((type.type == Type::hist || type.type == Type::lhist) &&
      (LINUX_VERSION_CODE >= KERNEL_VERSION(4, 6, 0)))
      map_type_ = BPF_MAP_TYPE_PERCPU_HASH;
  else
    map_type_ = BPF_MAP_TYPE_HASH;
```

in the `zero_map` function

```
  if ( ... map.type_.type == Type::lhist ... )
    value_size *= ncpus_;
```

The `ncpu` scaling is only valid for a `PERCPU` type as those have a
value for each CPU. On a kernel < 4.6.0 this will likely cause an issue.

This commit makes the type the map was created with available in the map
object. That way the actual map type can be used instead of guessing
based on value type.