Fix maptype bugs
The `ncpus_` scaling is only valid when the map is of type PER_CPU. This
isn't always the case as PER_CPU isn't used on kernels < 4.6.0.

Because 0 initialized vectors were used this didn't lead to bugs (afaik)
but it is still good to fix it.

This commit changes that by looking at the type with which the map was
created, instead of guessing it based on the type of the content.