Split lookup and initialization
If the specified key does not exist `CreateMapLookupElem` returns
the "zero-value" for the specified type. This makes it impossible to
know whether the value at this key actually was zero or that it did not
exist and the zero value was returned.

This zero value initialization is usually OK but it is an is an issue
for functions like `max` as the number might be smaller than 0.

This commit splits the function in two:
- CreateMapLookupElem, the call to the `map_lookup_elem` function,
making it easier for others to implement their own `no key` handling.
- CreateMapLookupElemValue, which does a lookup but returns a value
initialized to it's "zero-value" when the requested key did not exist.