Fix negative number handling of min/max
The `min` and `max` functions had some issues. `min` only supported a 32 bit
range and `max` only did positive numbers.

The most significant change in this commit is the switch from a per-cpu
to a "normal" hash. Per-cpu hashes are zero filled once the first key is
written, which makes it impossible to differentiate between an actual
zero and a "no value zero" making it impossible to `max` negative numbers.

The downside is that all the benefits of per-cpu hashes are lost now
too, making max less suited for the highly concurrent use cases.

An alternative could be to initialize the per_cpu hash ourselves with
an appropriate initial value based on the function and whether the value
is signed or not. But that means we have to do proper warning and
support type casting, which isn't possible yet. It would also add
complexity without a clear benefit or use case.
The `min` and `max` functions are not used in any of the examples at the
moment

Fixes #892 #983