Support wildcards in probe path (#879)
Fixes #688. One case that remains tricky is usdt:/my/app/* because path includes leading and trailing colons in the lexer. It works if you provide a probe name, though. Maybe this is fine for now.

I'm not sure whether all the tests are needed or wanted, but they were helpful for testing, so I left them in.