Fix map field access
javierhonduco reported inconsistencies when dereferencing
map that was assigned cast pointer.

Following code:

  #include <linux/mm.h>

  kprobe:do_wp_page
  {
    @a = (struct vm_fault*)arg0;
    printf("%llx %llx\n", @a->address, @a->address);
    printf("------\n");
  }

would return different values for same field, like:
  ffffc90006836000 1db70

The problem is that we don't distinguish if the assignment expression
is pointer, in which case we want to pass along only the pointer and
not read the whole cast struct.

The pointer checks were missing also in CreateMapLookupElem,
when we need to treat pointer cast map values as numbers and
not as values of their structs.

Fixes #930