Add script to compare tool codegen between builds
This script makes it easier to test for unexpected changes in the code
generated for our "real world" tools.

It takes two bpftrace builds (binary) and a directory containing tools
as arguments and prints the diff for each of the tools in the tool
directory, e.g:

```
$ sudo /vagrant/scripts/compare_tool_codegen.sh bpftrace bpftrace_master /vagrant/tools/
Checking bashreadline
Checking biolatency
Checking biosnoop
Change detected for script: biosnoop
--- a_biosnoop	2019-11-07 15:48:23.987667861 +0000
+++ b_biosnoop	2019-11-07 15:48:24.027668859 +0000
@@ -361,7 +361,8 @@
   %1 = bitcast i64* %"$now" to i8*
   call void @llvm.lifetime.start.p0i8(i64 -1, i8* %1)
   store i64 0, i64* %"$now"
-  %strcmp.result = alloca i64
+  %strcmp.char = alloca i8
+  %strcmp.result = alloca i8
   %lookup_elem_val21 = alloca [16 x i8]
```

It's intended to be used as an extra, manual, verification step to
make sure our tools aren't unexpectedly impacted by a change.