Unset `CMAKE_REQUIRED_LIBRARIES` to avoid influencing tests
On a fedora system without `binutils-devel` the build fails with
`pthread not found`:

```
-- Please install the libbpf development package (missing: LIBBPF_LIBRARIES LIBBPF_INCLUDE_DIRS)
-- Please install the libbfd development package (missing: LIBBFD_LIBRARIES LIBBFD_INCLUDE_DIRS)
-- Performing Test LIBBFD_DISASM_FOUR_ARGS_SIGNATURE
-- Performing Test LIBBFD_DISASM_FOUR_ARGS_SIGNATURE - Failed
-- Please install the libopcodes development package (missing: LIBOPCODES_LIBRARIES LIBOPCODES_INCLUDE_DIRS)
-- Looking for sys/sdt.h
-- Looking for sys/sdt.h - not found
-- Looking for pthread.h
-- Looking for pthread.h - not found
CMake Error at /usr/share/cmake/Modules/FindPackageHandleStandardArgs.cmake:137 (message):
  Could NOT find Threads (missing: Threads_FOUND)
```

This is  caused by the `CMAKE_REQUIRED_LIBRARIES` being
"leaked" between tests combined with CMP0075 . When testing for threads
cmake still links agains `bfd` and `opcodes` which do not exist and the
test fails. To avoid this we have to clear `CMAKE_REQUIRED_LIBRARIES`
after the test.