Implement `signal`
This wraps the `bpf_send_signal` bpf helper that allows you to send a
`signal(7)` to the current task.

To keep the code simple the return code from signal is currently
ignored, also making it impossible to use `signal` in an assignment. If
the need arises it should be relatively easy to provide access to the
return code.