Allow string literals as signal specifiers
This allows the use of common signal names, like SIGKILL or KILL,
instead of having to remember the right ID.
As IDs are arch specific (see `signal(7)`) the use of string literals
also improve script portability a bit.

The implementation differs in that it takes a string literal as
argument which is translated into an int during codegen.
This avoid introducing a set of global identifiers (like enums) that
can be used in places that do not make sense.