Add builtin: cpid
This adds the `cpid` builtin which makes the pid of the spawned child
(`-c command`) available to programs:

```
bpftrace -e '
tracepoint:syscalls:sys_enter_openat /pid == cpid/
{
  printf("%d :: %s\n", cpid, str(args->filename));
}' -c ls
22804 :: /etc/ld.so.cache
22804 :: /lib/x86_64-linux-gnu/libselinux.so.1
22804 :: /lib/x86_64-linux-gnu/libc.so.6
```

If no command is specified the program will not run:

```
stdin:1:46-50: ERROR: cpid cannot be used without child command
tracepoint:syscalls:sys_enter_openat /pid == cpid/ { printf("%s\n", str(args->filename)); }
```

The biggest change is that the `fork()` now happens before the codegen
phase as the pid has to be available at that time. To do this the
spawning of the child had to be taking out of bpftrace::run and moved
into a function callable from `main`.

Fixes: #298