Implement bitfield support
Bitfields are a pretty commonly used C feature (especially in kernel
land). Having bpftrace support would make bpftrace easier to use for
kernel developers.

Example script:
```

kprobe:ip_send_check{
  printf("version is %x, ihl is %x\n", ((struct iphdr*)arg0)->version, ((struct iphdr*)arg0)->ihl);
}
```

`struct iphdr*` is defined as:
```
struct iphdr {
	__u8	ihl:4,
		version:4;
        ...
```

Before, we'd get the following output:
```
Attaching 1 probe...
version is 45, ihl is 45
version is 45, ihl is 45
version is 45, ihl is 45
...
```

Now we get:
```
Attaching 1 probe...
version is 4, ihl is 5
version is 4, ihl is 5
version is 4, ihl is 5
version is 4, ihl is 5
...
```

This closes #969 .